new descriptive columns added to the fields table
Additional information columns to describe the transcribed observations were added to the "field" table, including the period of the observations (whether the observation was an instantaneous recording, a mean value, etc), the time of day of the observation if applicable, the measurement type (original observation, mean value, summary value), the original measurement unit, the SI measurement unit, and the measurement abbreviation. This is to work towards data output and validation purposes.